2_course/web/lab3/lab3_reinforcement_learning/rl-agent/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.9
dataclasses-json>=0.6.7
